Realesa second activity with edit text and save

+++ app/src/main/java/ru/netology/nmedia/DTO/Post.kt
@@ -1,14 +1,11 @@
 package ru.netology.nmedia.DTO
-
-import java.time.LocalDate
-
 data class Post(
-    val id: Long ,
+    val id: Long,
     val author: String,
     val publisher: String,
     val content: String,
     val countLiked: Long,
     val countShare: Long,
-    val countView: Long,
-    val likeBeMy: Boolean,
+    val countView: Int,
+    val likeBeMy: Boolean
 )
